+++ Stardust/src/scene/Scene.cpp
@@ -57,7 +57,7 @@ namespace stardust
 		return m_application.GetSceneManager();
 	}
 
-	f32 Scene::GetElapsedTime() const noexcept
+	f64 Scene::GetElapsedTime() const noexcept
 	{
 		return m_application.GetElapsedTime();
 	}
